more robust serialization and deserialization
instead of hard coded serialization and deserialization routines, each
serializable object declares a list of fields.

further, each field now has a key so that missing and extra fields can
be detected. when fields are missing, a default value callback supplies
the default value. when there are extra fields, they are ignored.

closes #16